+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add Intel RSD Network card information decoder #81
 - Add Intel RSD PCIe information decoder #82
 - Add Intel RSD TPM information decoder #85
+- Add Intel RSD TXT information decoder #86
 
 ### Fixed